Moved join() functions into a separate header
