+++ src/Utility/CMakeLists.txt
@@ -43,6 +43,7 @@ set(UTILITY_HEADERS
         Streams/OutputStream.h
         Streams/StringOutputStream.h
         String/Ascii.h
+        String/Join.h
         String/Split.h
         String/TransparentFunctors.h
         String/Transformations.h
@@ -75,7 +76,8 @@ if(OE_BUILD_TESTS)
             String/Tests/Transformations_ut.cpp
             String/Tests/TransparentFunctors_ut.cpp
             String/Tests/Ascii_ut.cpp
-            String/Tests/Split_ut.cpp)
+            String/Tests/Split_ut.cpp
+            String/Tests/Join_ut.cpp)
 
     add_library(test_utility OBJECT ${TEST_UTILITY_SOURCES})
     target_link_libraries(test_utility PUBLIC testing_unit utility)
